refactor: move testIncrement and testIncrementUpperBoundary to ByteUtilsTest
